refactor: Cart에서 사용되는 Component 위치 변경
